edit is_ series

+++ ft_memchr.c
@@ -2,10 +2,10 @@
 
 void *ft_memchr(const void *s, int c, size_t n)
 {
-    long unsigned int i;
+    size_t i;
     unsigned char *memory;
+    
     memory = (unsigned char *)s;
-
     i = 0;
     while (i < n)
     {
